Removed more unused code
